validate ownership before allow close/delete

+++ functions/post/deletePost.js
@@ -7,7 +7,12 @@ module.exports = async (req, res) => {
             res.status(400).send("document id not found")
             return
         }
-        var snap = await db.collection("posts").doc(id).delete()
+        const snap = await db.collection("posts").doc(id).get()
+        if (snap.data().d.uid !== req.authId) {
+            res.status(403).send("invalid ownership")
+            return
+        }
+        await db.collection("posts").doc(id).delete()
         res.send("deleted")
     } catch (err) {
         console.log(err)
